+++ app/Modules/Dashboard/Controllers/DashboardController.php
@@ -18,6 +18,7 @@ class DashboardController extends Controller
 {
     public function index()
     {
+
         return view('dashboard.index')
             ->with('widgets', DashboardWidgets::listsByOrder());
     }