+++ app/Modules/Employees/Models/Employee.php
@@ -28,5 +28,22 @@ class Employee extends Model
             ->where('resource_table','=','employees');
     }
 
+    //mutators
+    public function setFirstNameAttribute($value){
+        $this->attributes['first_name'] = ucfirst($value);
+    }
+
+    public function setLastNameAttribute($value){
+        $this->attributes['last_name'] = ucfirst($value);
+    }
+
+    public function setFullNameAttribute(){
+        $this->attributes['full_name'] = $this->attributes['first_name'] . ' ' . $this->attributes['last_name'];
+    }
+
+    public function setShortNameAttribute($value){
+        $this->attributes['short_name'] = $this->attributes['first_name'] . ' ' . substr($this->attributes['last_name'],0,1) . '.';
+    }
+
 
 }