+++ app/Modules/Quotes/Views/quotes/_js_create.blade.php
@@ -6,6 +6,7 @@
 
         $('#create-quote').on('shown.bs.modal', function () {
             $("#create_client_name").focus();
+            $("#create_client_name").val(clientName);
             $('#create_client_name').autocomplete({
                 appendTo: '#create-quote',
                 source: '{{ route('clients.ajax.lookup') }}',