+++ database/factories/ClientFactory.php
@@ -13,11 +13,14 @@ use Faker\Generator as Faker;
 |
 */
 
-$factory->define(\FI\Modules\Users\Models\User::class, function (Faker $faker) {
+$factory->define(\FI\Modules\Clients\Models\Client::class, function (Faker $faker) {
     return [
-        'name' => $faker->name,
+        'name' => $faker->company,
+        'address' => $faker->streetAddress,
+        'city' => $faker->city,
+        'state' => $faker->stateAbbr,
+        'zip' => $faker->postcode,
+        'phone' => $faker->numerify('(###) ###-####'),
         'email' => $faker->unique()->safeEmail,
-        'password' => '$2y$10$TKh8H1.PfQx37YgCzwiKb.KjNyWgaHb9cbcoQgdIVFlYg7B77UdFm', // secret
-        'remember_token' => str_random(10),
     ];
 });