+++ database/factories/ProductFactory.php
@@ -13,11 +13,15 @@ use Faker\Generator as Faker;
 |
 */
 
-$factory->define(\FI\Modules\Users\Models\User::class, function (Faker $faker) {
+$factory->define(\FI\Modules\Products\Models\Product::class, function (Faker $faker) {
     return [
-        'name' => $faker->name,
-        'email' => $faker->unique()->safeEmail,
-        'password' => '$2y$10$TKh8H1.PfQx37YgCzwiKb.KjNyWgaHb9cbcoQgdIVFlYg7B77UdFm', // secret
-        'remember_token' => str_random(10),
+        'name' => ucfirst($faker->word),
+        'description' => $faker->sentence,
+        'serialnum' => $faker->randomNumber(9),
+        'active' => 1,
+        'cost' => $faker->randomFloat(2, 19, 150),
+        'category' => 'Product',
+        'type' => 'Product Type',
+        'numstock' => $faker->numberBetween(1,10),
     ];
 });